requirements.txt
alembic==1.9.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
